requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy 
databases
# psycopg2-binary
pymysql
python-dotenv